优化

+++ Sub-Store/sub-store.sh
@@ -14,7 +14,9 @@ cnpm install
 
 echo -e "开始初始化接口\n"
 cp -r /git/backend /Sub-Store/
-cp -r /git/nginx /Sub-Store/
+if [ ! -f "$nginx/fron.conf"];then
+    cp -r /git/nginx /Sub-Store/
+fi
 cd "$backend" 
 cnpm install
